type gaurds and declaration files
