+++ core/dts/type.vno.d.ts
@@ -1,26 +1,28 @@
 export {
+  CompId,
+  Desc,
+  IdxOf,
+  Join,
+  ParserMethod,
+  Pattern,
+  ResolveAttrs,
+  ResolveSrc,
+  Split,
+  Src,
+  Tag,
+  TsHelper,
+} from "./vno.utility.d.ts";
+
+export {
   Component,
   ComponentContainer,
   ComponentList,
   ComponentType,
+  DepsList,
+  Options,
   ParsedData,
   ParseSelf,
+  Queue,
   RawData,
-} from "./vno.component.d.ts";
-
-export {
-  CompId,
-  IdxOf,
-  Join,
-  Pattern,
-  Split,
-  TsHelper,
-} from "./vno.utility.d.ts";
-
-export { ParserMethod } from "./vno.parser.d.ts";
-
-export { Desc, Src, Tag } from "./vno.analysis.d.ts";
-
-export { DepsList, Options, Queue, Storage } from "./vno.factory.d.ts";
-
-export { ResolveAttrs, ResolveSrc } from "./vno.resolver.d.ts";
+  Storage,
+} from "./vno.factory.d.ts";
